+++ cate/ops/aggregate.py
@@ -41,7 +41,7 @@ from cate.core.types import VarNamesLike, DatasetLike, ValidationError, DimNames
 from cate.ops.normalize import adjust_temporal_attrs
 
 
-@op(tags=['aggregate', 'temporal'], version='1.0')
+@op(tags=['aggregate', 'temporal'], version='1.5')
 @op_input('ds', data_type=DatasetLike)
 @op_input('var', value_set_source='ds', data_type=VarNamesLike)
 @op_return(add_history=True)
@@ -49,15 +49,21 @@ def long_term_average(ds: DatasetLike.TYPE,
                       var: VarNamesLike.TYPE = None,
                       monitor: Monitor = Monitor.NONE) -> xr.Dataset:
     """
-    Perform long term average of the given dataset by doing a mean of monthly
-    values over the time range covered by the dataset. E.g. it averages all
-    January values, all February values, etc, to create a dataset with twelve
-    time slices each containing a mean of respective monthly values.
+    Create a 'mean over years' dataset by averaging the values of the given input
+    dataset over all years. The output is a climatological dataset with the same
+    resolution as the input dataset. E.g. a daily input dataset will create a daily
+    climatology consisting of 365 days, a monthly input dataset will create a monthly
+    climatology, etc.
+
+    Seasonal input datasets must have matching seasons over all years denoted by the
+    same date each year. E.g., first date of each quarter. The output dataset will
+    then be a seasonal climatology where each season is denoted with the same date
+    as in the input dataset.
 
     For further information on climatological datasets, see
     http://cfconventions.org/cf-conventions/v1.6.0/cf-conventions.html#climatological-statistics
 
-    :param ds: A monthly dataset to average
+    :param ds: A dataset to average
     :param var: If given, only these variables will be preserved in the resulting dataset
     :param monitor: A progress monitor
     :return: A climatological long term average dataset
@@ -70,12 +76,8 @@ def long_term_average(ds: DatasetLike.TYPE,
                               ' {}. Running the normalize operation on this'
                               ' dataset may help'.format(ds.time.dtype))
 
-    # Check if we have a monthly dataset
     try:
-        if ds.attrs['time_coverage_resolution'] != 'P1M':
-            raise ValidationError('Long term average operation expects a monthly dataset'
-                                  ' running temporal aggregation on this dataset'
-                                  ' beforehand may help.')
+        t_resolution = ds.attrs['time_coverage_resolution']
     except KeyError:
         raise ValidationError('Could not determine temporal resolution. Running'
                               ' the adjust_temporal_attrs operation beforehand may'
@@ -87,10 +89,26 @@ def long_term_average(ds: DatasetLike.TYPE,
     if var:
         retset = select_var(retset, var)
 
+    if t_resolution == 'P1D':
+        return _lta_daily(retset, monitor)
+    elif t_resolution == 'P1M':
+        return _lta_monthly(retset, monitor)
+    else:
+        return _lta_general(retset, monitor)
+
+
+def _lta_monthly(ds: xr.Dataset, monitor: Monitor):
+    """
+    Carry out a long term average on a monthly dataset
+
+    :param ds: Dataset to aggregate
+    :param monitor: Progress monitor
+    :return: Aggregated dataset
+    """
     time_min = pd.Timestamp(ds.time.values[0])
     time_max = pd.Timestamp(ds.time.values[-1])
-
     total_work = 100
+    retset = ds
 
     with monitor.starting('LTA', total_work=total_work):
         monitor.progress(work=0)
@@ -122,6 +140,176 @@ def long_term_average(ds: DatasetLike.TYPE,
     return retset
 
 
+def _groupby_day(ds: xr.Dataset, monitor: Monitor, step: float):
+    """
+    Groupby the given dataset by day of month and apply mean to it
+
+    :param ds: Dataset to aggregate
+    :param monitor: Progress monitor
+    :param step: Progress step
+    """
+    kwargs = {'monitor': monitor, 'step': step}
+    return ds.groupby('time.day', squeeze=False).apply(_mean, **kwargs)
+
+
+def _lta_daily(ds: xr.Dataset, monitor: Monitor):
+    """
+    Carry out a long term average of a daily dataset
+
+    :param ds: Dataset to aggregate
+    :param monitor: Progress monitor
+    :return: Aggregated dataset
+    """
+    time_min = pd.Timestamp(ds.time.values[0])
+    time_max = pd.Timestamp(ds.time.values[-1])
+    total_work = 100
+    retset = ds
+
+    with monitor.starting('LTA', total_work=total_work):
+        monitor.progress(work=0)
+        step = total_work / 366
+        kwargs = {'monitor': monitor, 'step': step}
+        retset = retset.groupby('time.month', squeeze=False).apply(_groupby_day, **kwargs)
+
+    # Make the return dataset CF compliant
+    retset = retset.stack(time=('month', 'day'))
+
+    # Get rid of redundant dates
+    drop = [(2, 29), (2, 30), (2, 31), (4, 31), (6, 31),
+            (9, 31), (11, 31)]
+    retset = retset.drop(drop, dim='time')
+
+    # Turn month, day coordinates to time
+    retset = retset.reset_index('time')
+    retset = retset.drop(['month', 'day'])
+    time_coord = pd.date_range(start='{}-01-01'.format(time_min.year),
+                               end='{}-12-31'.format(time_min.year),
+                               freq='D')
+    if len(time_coord) == 366:
+        time_coord = time_coord.drop(np.datetime64('{}-02-29'.format(time_min.year)))
+    retset['time'] = time_coord
+
+    climatology_bounds = xr.DataArray(data=np.tile([time_min, time_max],
+                                                   (365, 1)),
+                                      dims=['time', 'nv'],
+                                      name='climatology_bounds')
+    retset['climatology_bounds'] = climatology_bounds
+    retset.time.attrs = ds.time.attrs
+    retset.time.attrs['climatology'] = 'climatology_bounds'
+
+    for var in retset.data_vars:
+        try:
+            retset[var].attrs['cell_methods'] = \
+                retset[var].attrs['cell_methods'] + ' time: mean over years'
+        except KeyError:
+            retset[var].attrs['cell_methods'] = 'time: mean over years'
+
+    return retset
+
+
+def _lta_general(ds: xr.Dataset, monitor: Monitor):
+    """
+    Try to carry out a long term average in a general case, notably
+    in the case of having seasonal datasets
+
+    :param ds: Dataset to aggregate
+    :param monitor: Progress monitor
+    :return: Aggregated dataset
+    """
+    time_min = pd.Timestamp(ds.time.values[0])
+    time_max = pd.Timestamp(ds.time.values[-1])
+    total_work = 100
+    retset = ds
+
+    # The dataset should feature time periods consistent over years
+    # and denoted with the same dates each year
+    if not _is_seasonal(ds.time):
+        raise ValidationError("A long term average dataset can not be created for"
+                              " a dataset with inconsistent seasons.")
+
+    # Get 'representative year'
+    c = 0
+    for group in ds.time.groupby('time.year'):
+        c = c + 1
+        if c == 1:
+            rep_year = group[1].time
+            continue
+        if c == 2 and len(group[1].time) > len(rep_year):
+            rep_year = group[1].time
+            break
+
+    with monitor.starting('LTA', total_work=total_work):
+        monitor.progress(work=0)
+        step = total_work / len(rep_year.time)
+        kwargs = {'monitor': monitor, 'step': step}
+        retset = retset.groupby('time.month', squeeze=False).apply(_groupby_day, **kwargs)
+
+    # Make the return dataset CF compliant
+    retset = retset.stack(time=('month', 'day'))
+
+    # Turn month, day coordinates to time
+    retset = retset.reset_index('time')
+    retset = retset.drop(['month', 'day'])
+    retset['time'] = rep_year.time
+
+    climatology_bounds = xr.DataArray(data=np.tile([time_min, time_max],
+                                                   (len(rep_year), 1)),
+                                      dims=['time', 'nv'],
+                                      name='climatology_bounds')
+    retset['climatology_bounds'] = climatology_bounds
+    retset.time.attrs = ds.time.attrs
+    retset.time.attrs['climatology'] = 'climatology_bounds'
+
+    for var in retset.data_vars:
+        try:
+            retset[var].attrs['cell_methods'] = \
+                retset[var].attrs['cell_methods'] + ' time: mean over years'
+        except KeyError:
+            retset[var].attrs['cell_methods'] = 'time: mean over years'
+
+    return retset
+
+
+def _is_seasonal(time: xr.DataArray):
+    """
+    Check if the given timestamp dataarray features consistent
+    seasons. E.g. Each year has the same date-month values in it.
+    """
+    c = 0
+    for group in time.groupby('time.year'):
+        # Test (month, day) dates of all years against
+        # (month, day) dates of the first year, or second
+        # year in case the first year is not full
+        c = c + 1
+        np_time = group[1].time.values
+        months = pd.DatetimeIndex(np_time).month
+        days = pd.DatetimeIndex(np_time).day
+        if c == 1:
+            first_months = months
+            first_days = days
+            continue
+        elif c == 2:
+            second_months = months
+            second_days = days
+            if len(second_months) > len(first_months):
+                test = zip(second_months, second_days)
+                for date in zip(first_months, first_days):
+                    if date not in test:
+                        return False
+            else:
+                test = zip(first_months, first_days)
+                for date in zip(second_months, second_days):
+                    if date not in test:
+                        return False
+            continue
+
+        for date in zip(months, days):
+            if date not in test:
+                return False
+
+    return True
+
+
 def _mean(ds: xr.Dataset, monitor: Monitor, step: float):
     """
     Calculate mean of the given dataset and update the given monitor.
@@ -147,7 +335,7 @@ def temporal_aggregation(ds: DatasetLike.TYPE,
                          custom_resolution: str = None,
                          monitor: Monitor = Monitor.NONE) -> xr.Dataset:
     """
-    Perform monthly aggregation of dataset according to the given
+    Perform aggregation of dataset according to the given
     method and output resolution.
 
     Note that the operation does not perform weighting. Depending on the